Funny Fractal Encryption Pre-Alpha Renderer
A simple fractal render program that generates a PPM. Fwiw, here is some pure c99 code. ** Be warned **, the code creates an output file called ct_ffe_cipher.ppm. Can you run and open the file in a graphics editor? It works for me with the Gimp. This non-sense with the hardcoded file name will change to gain data in the arguments from the command line in my next iteration of the code. Sorry about the crude state. ;^o 

to compile in GCC:
gcc -Wall -o ffe --std=c99 -lm ffe.c

Read all of for further context:

https://groups.google.com/d/topic/sci.crypt/xytM7aFRfjQ/discussion